vis2b: fix possibly leak/corruption
when sending ubx types via lock free fifos, a copy is made. This copy
considers the array length, but does not follow any pointers in the
data-structure.

When the vis2b block reads into its local instance of gc_pose, two
things happen: firstly, the local gc_pose's pointers are overwritten,
thus the references to that memory are lost. Secondly, when vis2b is
shutdown and tries to free it's rotation and translation buffers, it
will likely crash (because that memory has been freed already (by it's
owner, the upstream block) or is invalid).

The quick solution is to just allocate a pose array without allocating
the contained rotation and translation buffers (and not to free these
either). After a successful read (and only then), these fields will be
set and the nbx can be triggered.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>